+++ pkg/resource/resource_goal.go
@@ -36,7 +36,7 @@ type Goal struct {
 	AdditionalSecretOutputs []PropertyKey         // outputs that should always be treated as secrets.
 	Aliases                 []URN                 // additional URNs that should be aliased to this resource.
 	ID                      ID                    // the expected ID of the resource, if any.
-	CustomTimeouts          *CustomTimeouts       // an optional config object for resource options
+	CustomTimeouts          CustomTimeouts        // an optional config object for resource options
 }
 
 // NewGoal allocates a new resource goal state.
@@ -45,7 +45,7 @@ func NewGoal(t tokens.Type, name tokens.QName, custom bool, props PropertyMap,
 	propertyDependencies map[PropertyKey][]URN, deleteBeforeReplace bool, ignoreChanges []string,
 	additionalSecretOutputs []PropertyKey, aliases []URN, id ID, customTimeouts *CustomTimeouts) *Goal {
 
-	return &Goal{
+	g := &Goal{
 		Type:                    t,
 		Name:                    name,
 		Custom:                  custom,
@@ -61,6 +61,11 @@ func NewGoal(t tokens.Type, name tokens.QName, custom bool, props PropertyMap,
 		AdditionalSecretOutputs: additionalSecretOutputs,
 		Aliases:                 aliases,
 		ID:                      id,
-		CustomTimeouts:          customTimeouts,
 	}
+
+	if customTimeouts != nil {
+		g.CustomTimeouts = *customTimeouts
+	}
+
+	return g
 }