+++ CHANGELOG.md
@@ -15,18 +15,19 @@ CHANGELOG
   [#2871](https://github.com/pulumi/pulumi/issues/2871))
 - Promises/Outputs can now be directly exported as the top-level (i.e. not-named) output of a Stack.
   (fixes [#2910](https://github.com/pulumi/pulumi/issues/2910))
+  
 - Add support for importing existing resources to be managed using Pulumi. A resource can be imported
   by setting the `import` property in the resource options bag when instantiating a resource. In order to
   successfully import a resource, its desired configuration (i.e. its inputs) must not differ from its
   actual configuration (i.e. its state) as calculated by the resource's provider.
+  
+- Add the ability to pass a customTimeouts object from the providers across the engine
+  to resource management. (fixes [#2665](https://github.com/pulumi/pulumi/issues/2665))
 
 ## 0.17.22 (2019-07-11)
 
 - Improve update performance in cases where a large number of log messages are
   reported during an update.
-
-- Add the ability to pass a customTimeouts object from the providers across the engine
-  to resource management. (fixes [#2665](https://github.com/pulumi/pulumi/issues/2665))
 
 ## 0.17.21 (2019-06-26)